apps/api/pyproject.toml
[project]
name = "wan-api"
version = "2.1.0"
description = "FastAPI backend for Wan 2.1 video generation"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "replicate>=0.34.1",
    "supabase>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
